this is not working and i have no idea what im doing
