--- Notyet.java
@@ -1,22 +1,0 @@
-import greenfoot.*;  // (World, Actor, GreenfootImage, Greenfoot and MouseInfo)
-
-/**
- * Write a description of class Notyet here.
- * 
- * @author (your name) 
- * @version (a version number or a date)
- */
-public class Notyet extends Actor
-{
-    /**
-     * Act - do whatever the Notyet wants to do. This method is called whenever
-     * the 'Act' or 'Run' button gets pressed in the environment.
-     */
-    public void act()
-    {
-        // Add your action code here.
-        if (Greenfoot.mouseClicked(this)){
-            System.out.println("Too early!");
-        }
-    }
-}
