--- Now.ctxt
@@ -1,13 +1,0 @@
-#BlueJ class context
-comment0.target=Now
-comment0.text=\r\n\ Write\ a\ description\ of\ class\ Now\ here.\r\n\ \r\n\ @author\ (your\ name)\ \r\n\ @version\ (a\ version\ number\ or\ a\ date)\r\n
-comment1.params=
-comment1.target=Now()
-comment1.text=\r\n\ Act\ -\ do\ whatever\ the\ Now\ wants\ to\ do.\ This\ method\ is\ called\ whenever\r\n\ the\ 'Act'\ or\ 'Run'\ button\ gets\ pressed\ in\ the\ environment.\r\n
-comment2.params=
-comment2.target=void\ onClick()
-comment3.params=
-comment3.target=void\ act()
-comment4.params=
-comment4.target=int\ getTime()
-numComments=5
