--- ReactionGame.ctxt
@@ -1,8 +1,0 @@
-#BlueJ class context
-comment0.target=ReactionGame
-comment0.text=\r\n\ Write\ a\ description\ of\ class\ ReactionGame\ here.\r\n\ \r\n\ @author\ (your\ name)\ \r\n\ @version\ (a\ version\ number\ or\ a\ date)\r\n
-comment1.params=
-comment1.target=ReactionGame()
-comment2.params=
-comment2.target=void\ act()
-numComments=3
