+++ ReactionTimeGame.ctxt
@@ -1,0 +1,9 @@
+#BlueJ class context
+comment0.target=ReactionTimeGame
+comment1.params=
+comment1.target=ReactionTimeGame()
+comment2.params=
+comment2.target=void\ onClick()
+comment3.params=
+comment3.target=void\ act()
+numComments=4
