+++ RedButton.ctxt
@@ -1,0 +1,9 @@
+#BlueJ class context
+comment0.target=RedButton
+comment0.text=\r\n\ Write\ a\ description\ of\ class\ RedButton\ here.\r\n\ \r\n\ @author\ (your\ name)\ \r\n\ @version\ (a\ version\ number\ or\ a\ date)\r\n
+comment1.params=
+comment1.target=void\ act()
+comment1.text=\r\n\ Act\ -\ do\ whatever\ the\ RedButton\ wants\ to\ do.\ This\ method\ is\ called\ whenever\r\n\ the\ 'Act'\ or\ 'Run'\ button\ gets\ pressed\ in\ the\ environment.\r\n
+comment2.params=
+comment2.target=boolean\ getEarly()
+numComments=3
