--- Results.ctxt
@@ -1,7 +1,0 @@
-#BlueJ class context
-comment0.target=Results
-comment0.text=\r\n\ Write\ a\ description\ of\ class\ Results\ here.\r\n\ \r\n\ @author\ (your\ name)\ \r\n\ @version\ (a\ version\ number\ or\ a\ date)\r\n
-comment1.params=
-comment1.target=Results()
-comment1.text=\r\n\ Constructor\ for\ objects\ of\ class\ Results.\r\n\ \r\n
-numComments=2
