--- Results.java
@@ -1,21 +1,0 @@
-import greenfoot.*;  // (World, Actor, GreenfootImage, Greenfoot and MouseInfo)
-
-/**
- * Write a description of class Results here.
- * 
- * @author (your name) 
- * @version (a version number or a date)
- */
-public class Results extends ReactionGame
-{
-
-    /**
-     * Constructor for objects of class Results.
-     * 
-     */
-    public Results()
-    {
-        Label delay = new Label(Now.getTime() + " ms", 50);
-        
-    }
-}
