+++ project.greenfoot
@@ -2,42 +2,34 @@
 class.Back.image=back.png
 class.Button.image=button.png
 class.MyWorld.image=graybackground.png
-class.Notyet.image=red.png
-class.Now.image=now.png
 class.Reaction.image=reaction.png
-class.ReactionGame.image=1graybackground.png
-class.Results.image=graybackground.png
-dependency1.from=ReactionGame
-dependency1.to=Notyet
+class.ReactionTimeGame.image=1graybackground.png
+class.RedButton.image=red.png
+dependency1.from=ReactionTimeGame
+dependency1.to=RedButton
 dependency1.type=UsesDependency
-dependency2.from=ReactionGame
-dependency2.to=Back
+dependency2.from=MyWorld
+dependency2.to=Reaction
 dependency2.type=UsesDependency
-dependency3.from=ReactionGame
-dependency3.to=Now
+dependency3.from=MyWorld
+dependency3.to=Button
 dependency3.type=UsesDependency
 dependency4.from=MyWorld
-dependency4.to=Reaction
+dependency4.to=Label
 dependency4.type=UsesDependency
-dependency5.from=MyWorld
-dependency5.to=Button
+dependency5.from=Reaction
+dependency5.to=ReactionTimeGame
 dependency5.type=UsesDependency
-dependency6.from=MyWorld
-dependency6.to=Label
+dependency6.from=Back
+dependency6.to=MyWorld
 dependency6.type=UsesDependency
-dependency7.from=Reaction
-dependency7.to=ReactionGame
-dependency7.type=UsesDependency
-dependency8.from=Back
-dependency8.to=MyWorld
-dependency8.type=UsesDependency
 editor.fx.0.height=737
 editor.fx.0.width=814
 editor.fx.0.x=360
 editor.fx.0.y=28
 height=838
-package.numDependencies=8
-package.numTargets=9
+package.numDependencies=6
+package.numTargets=7
 project.charset=UTF-8
 publish.hasSource=false
 publish.locked=true
@@ -60,61 +52,47 @@ target1.width=120
 target1.x=0
 target1.y=0
 target2.height=70
-target2.name=Results
+target2.name=ReactionTimeGame
 target2.showInterface=false
 target2.type=ClassTarget
-target2.width=120
+target2.width=130
 target2.x=0
 target2.y=0
 target3.height=70
-target3.name=ReactionGame
+target3.name=Back
 target3.showInterface=false
 target3.type=ClassTarget
 target3.width=120
 target3.x=0
 target3.y=0
 target4.height=70
-target4.name=Now
+target4.name=Label
 target4.showInterface=false
 target4.type=ClassTarget
 target4.width=120
 target4.x=0
 target4.y=0
 target5.height=70
-target5.name=Back
+target5.name=RedButton
 target5.showInterface=false
 target5.type=ClassTarget
 target5.width=120
 target5.x=0
 target5.y=0
 target6.height=70
-target6.name=Label
+target6.name=Reaction
 target6.showInterface=false
 target6.type=ClassTarget
 target6.width=120
 target6.x=0
 target6.y=0
 target7.height=70
-target7.name=Notyet
+target7.name=MyWorld
 target7.showInterface=false
 target7.type=ClassTarget
 target7.width=120
 target7.x=0
 target7.y=0
-target8.height=70
-target8.name=Reaction
-target8.showInterface=false
-target8.type=ClassTarget
-target8.width=120
-target8.x=0
-target8.y=0
-target9.height=70
-target9.name=MyWorld
-target9.showInterface=false
-target9.type=ClassTarget
-target9.width=120
-target9.x=0
-target9.y=0
 version=3.1.0
 width=1550
 world.lastInstantiated=MyWorld
